requirements.txt
openai==1.35.10
python-dotenv==1.0.1
websockets==10.4
asyncio==3.4.3
web3==6.20.0
flask